% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Erdos-Renyi}
\alias{Erdos-Renyi}
\alias{rgraph_er_gnm}
\alias{rgraph_er_gnp}
\title{Erdős-Renyi}
\usage{
rgraph_er_gnm(n, m, loop = FALSE)

rgraph_er_gnp(n, p, loop = FALSE)
}
\arguments{
\item{n}{Size.}

\item{m}{Number of edges in the \code{G(n,m)} model.}

\item{loop}{Logical. When \code{TRUE}, no loops.}

\item{p}{Probability as in the \code{G(n, p)} model.}
}
\description{
Erdős-Renyi
}
