% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rgraph_ba}
\alias{rgraph_ba}
\title{Barabasi-Albert}
\usage{
rgraph_ba(t, m, alpha = 1, a = 1, loop = FALSE)
}
\arguments{
\item{t}{Time steps.}

\item{m}{Number of ties added per node.}

\item{alpha}{Degree}

\item{a}{Baseline probability}

\item{loop}{Logical. When \code{TRUE} it allows for auto-links (loops).}
}
\description{
Barabasi-Albert
}
